pyproject.toml
[tool.poetry]
name = "fast-dw"
version = "0.1.0"
description = "A fast way to deploy a data warehouse, using dbt and other tools"
authors = ["João Tedeschi"]
license = "MIT"
readme = "README.md"
packages = [{include = "fast_dw"}]

[tool.poetry.dependencies]
python = "^3.10"
dbt-core = "^1.6.3"
dbt-bigquery = "^1.6.5"
faker = "^19.6.1"
pandas = "^2.1.0"
pyarrow = "^13.0.0"
google-cloud-storage = "^2.11.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
google-cloud-bigquery = "^3.11.4"
google-auth = "^2.23.0"
pandas-gbq = "^0.19.2"
pytest = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
